requirements.txt
quart~=0.18.3
uvicorn~=0.21.1
click~=8.1.3
aiofiles~=23.1.0
python-magic~=0.4.27